A proof-of-concept for defmt *without* interning strings.
This is just an experiment for now.

If the 'no-interning' feature is set, instead of

#[export_name = "{JSON...}"]
static S: u8 = 0;
&S as *const _ as usize

You get:

static S: &'static str = "{JSON...}";
&S as *const _ as usize

Note that I had to change the size of the interned ID from u16 to usize, because with no-interning it's now a pointer to actual data.